# pkg/master/api_v3/openapi.yaml
openapi: 3.0.3
info:
  title: PandaFuzz API
  description: |
    PandaFuzz is a distributed fuzzing orchestration platform that manages fuzzing bots,
    jobs, campaigns, and corpus management. This API provides comprehensive access to
    all platform features with versioning support.
  version: 3.0.0
  contact:
    name: PandaFuzz Team
    url: https://github.com/ethpandaops/pandafuzz
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api/v3
    description: Local development server
  - url: https://pandafuzz.example.com/api/v3
    description: Production server
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /corpus/sync:
    post:
      tags: [corpus]
      summary: Synchronize corpus between campaigns/jobs
      operationId: syncCorpus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorpusSyncRequest'
      responses:
        '200':
          description: Sync completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorpusSyncResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /corpus/promote:
    post:
      tags: [corpus]
      summary: Promote crash to corpus
      operationId: promoteCrashToCorpus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorpusPromotionRequest'
      responses:
        '201':
          description: Crash promoted to corpus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorpusFile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Crash Management
...
  /crashes/{crashId}:
    parameters:
      - $ref: '#/components/parameters/crashIdParam'
    
    get:
      tags: [crashes]
      summary: Get crash details
      operationId: getCrash
      responses:
        '200':
          description: Crash details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrashResult'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    crashIdParam:
      name: crashId
      in: path
      required: true
      description: Crash identifier
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CorpusFile:
      type: object
      required:
        - id
        - campaign_id
        - job_id
        - bot_id
        - filename
        - hash
        - size
        - coverage
        - new_coverage
        - parent_hash
        - generation
        - created_at
        - is_seed
      properties:
        id:
          type: string
          format: uuid
        campaign_id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        bot_id:
          type: string
          format: uuid
        filename:
          type: string
        hash:
          type: string
        size:
          type: integer
          format: int64
        coverage:
          type: integer
          format: int64
        new_coverage:
          type: integer
          format: int64
        parent_hash:
          type: string
        generation:
          type: integer
        created_at:
          type: string
          format: date-time
        synced_at:
          type: string
          format: date-time
          nullable: true
        is_seed:
          type: boolean
...
    CorpusSyncRequest:
      type: object
      required:
        - source_campaign_id
        - target_campaign_id
      properties:
        source_campaign_id:
          type: string
          format: uuid
        target_campaign_id:
          type: string
          format: uuid
        filter_coverage:
          type: boolean
        min_coverage:
          type: integer
          format: int64
    CorpusSyncResponse:
      type: object
      required:
        - synced_files
        - total_size
        - duration
      properties:
        synced_files:
          type: integer
        total_size:
          type: integer
          format: int64
        duration:
          type: string
    CorpusPromotionRequest:
      type: object
      required:
        - crash_id
        - campaign_id
        - reason
      properties:
        crash_id:
          type: string
          format: uuid
        campaign_id:
          type: string
          format: uuid
        reason:
          type: string
        priority:
          type: integer
          minimum: 1
          maximum: 10
...
    CrashResult:
      type: object
      required:
        - job_id
        - bot_id
        - hash
        - file_path
        - type
        - timestamp
        - size
      properties:
        id:
          type: string
          format: uuid
        job_id:
          type: string
          format: uuid
        bot_id:
          type: string
          format: uuid
        campaign_id:
          type: string
          format: uuid
        hash:
          type: string
        file_path:
          type: string
        type:
          type: string
          enum: [segfault, assertion, timeout, other]
        signal:
          type: integer
        exit_code:
          type: integer
        timestamp:
          type: string
          format: date-time
        size:
          type: integer
          format: int64
        is_unique:
          type: boolean
        input_base64:
          type: string
          format: byte
        output:
          type: string
        stack_trace:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...